requirements.txt
# Twitter Financial Signal Monitor

fastapi==0.110.0
uvicorn[standard]==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0
apscheduler==3.10.4
google-generativeai==0.3.2
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
typer==0.9.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
